Adding controller validation and logic, refactored services to inherit DbContext

+++ GameSource.API/Controllers/DeveloperController.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using GameSource.Models;
 using GameSource.Models.Enums;
@@ -27,7 +26,6 @@ namespace GameSource.API.Controllers.GameSource
         public async Task<ApiResponse> GetAll()
         {
             IEnumerable<Developer> result = await developerService.GetAllAsync();
-
             if (result == null)
                 return new ApiResponse(result, ResponseStatusCode.Error, "Could not return Developer list.", result.Count());
 
@@ -38,7 +36,6 @@ namespace GameSource.API.Controllers.GameSource
         public async Task<ApiResponse> GetByID(int id)
         {
             var result = await developerService.GetByIDAsync(id);
-
             if (result == null)
                 return new ApiResponse(result, ResponseStatusCode.Error, "Could not return a Developer.");
 
@@ -49,33 +46,40 @@ namespace GameSource.API.Controllers.GameSource
         public async Task<ApiResponse> Insert([FromBody] Developer developer)
         {
             int rows = await developerService.InsertAsync(developer);
-
             if (rows <= 0)
-                return new ApiResponse(rows, ResponseStatusCode.Error, "Could not create a Developer.", rows);
+                return new ApiResponse(ResponseStatusCode.Error, "Could not create a Developer.", rows);
 
-            return new ApiResponse(rows, ResponseStatusCode.Success, "Successfully created a new Developer.", rows);
+            return new ApiResponse(ResponseStatusCode.Success, "Successfully created a new Developer.", rows);
         }
 
-        [HttpPut]
-        public async Task<ApiResponse> Update([FromBody] Developer developer)
+        [HttpPut("{id}")]
+        public async Task<ApiResponse> Update(int id, [FromBody] Developer developer)
         {
-            int rows = await developerService.UpdateAsync(developer);
+            Developer updatedDeveloper = developerService.GetByID(id);
+            if (updatedDeveloper == null)
+                return new ApiResponse(ResponseStatusCode.Error, "Developer was not found. Please check the ID.");
 
+            updatedDeveloper.Name = developer.Name;
+            updatedDeveloper.Games = developer.Games;
+
+            int rows = await developerService.UpdateAsync(updatedDeveloper);
             if (rows <= 0)
-                return new ApiResponse(rows, ResponseStatusCode.Error, "Could not update Developer.", rows);
+                return new ApiResponse(ResponseStatusCode.Error, "Could not update Developer.", rows);
 
-            return new ApiResponse(rows, ResponseStatusCode.Success, "Successfully updated Developer.", rows);
+            return new ApiResponse(ResponseStatusCode.Success, "Successfully updated Developer.", rows);
         }
 
         [HttpDelete("{id}")]
         public async Task<ApiResponse> Delete(int id)
         {
+            if (id == 0)
+                return new ApiResponse(ResponseStatusCode.Error, "Developer was not found. Please check the ID.");
+
             int rows = await developerService.DeleteAsync(id);
-
             if (rows <= 0)
-                return new ApiResponse(rows, ResponseStatusCode.Error, "Could not delete Developer.", rows);
+                return new ApiResponse(ResponseStatusCode.Error, "Could not delete Developer.", rows);
 
-            return new ApiResponse(rows, ResponseStatusCode.Success, "Successfully deleted Developer.", rows);
+            return new ApiResponse(ResponseStatusCode.Success, "Successfully deleted Developer.", rows);
         }
     }
 }
